+++ src/magnetic.h
@@ -14,7 +14,7 @@
 int mgPutEv(uint8_t event);
 
 int magGetStat(void);
-uint32_t getFreq(void);
+uint16_t getFreq(void);
 void magnetic(void);
 
 #endif /* MAGNETIC_H_ */
